+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/ToggleMinOppgaveliste.tsx
@@ -4,10 +4,13 @@ import { InboxIcon, PersonIcon } from '@navikt/aksel-icons'
 import styled from 'styled-components'
 import { useAppSelector } from '~store/Store'
 import { Container } from '~shared/styled'
-import { Tilgangsmelding } from '~components/oppgavebenk/Tilgangsmelding'
+import { Tilgangsmelding } from '~components/oppgavebenk/components/Tilgangsmelding'
 import { useLocation, useNavigate } from 'react-router-dom'
-import { Filter, minOppgavelisteFiltre } from '~components/oppgavebenk/filter/oppgavelistafiltre'
-import { hentFilterFraLocalStorage, leggFilterILocalStorage } from '~components/oppgavebenk/filter/filterLocalStorage'
+import { Filter, minOppgavelisteFiltre } from '~components/oppgavebenk/oppgaveFiltrering/oppgavelistafiltre'
+import {
+  hentFilterFraLocalStorage,
+  leggFilterILocalStorage,
+} from '~components/oppgavebenk/oppgaveFiltrering/filterLocalStorage'
 import { useApiCall } from '~shared/hooks/useApiCall'
 import { hentGosysOppgaver, hentOppgaverMedStatus, OppgaveDTO, saksbehandlereIEnhetApi } from '~shared/api/oppgaver'
 import { isSuccess } from '~shared/api/apiUtils'
@@ -15,9 +18,9 @@ import {
   finnOgOppdaterSaksbehandlerTildeling,
   leggTilOppgavenIMinliste,
   sorterOppgaverEtterOpprettet,
-} from '~components/oppgavebenk/oppgaveutils'
+} from '~components/oppgavebenk/utils/oppgaveutils'
 import { MinOppgaveliste } from '~components/oppgavebenk/MinOppgaveliste'
-import { OppgavelistaWrapper } from '~components/oppgavebenk/OppgavelistaWrapper'
+import { Oppgavelista } from '~components/oppgavebenk/Oppgavelista'
 import { Saksbehandler } from '~shared/types/saksbehandler'
 
 type OppgavelisteToggle = 'Oppgavelista' | 'MinOppgaveliste'
@@ -27,20 +30,24 @@ export const ToggleMinOppgaveliste = () => {
   if (!innloggetSaksbehandler.skriveTilgang) {
     return <Tilgangsmelding />
   }
+
   const [oppgaveListeValg, setOppgaveListeValg] = useState<OppgavelisteToggle>('Oppgavelista')
-  const [, hentSaksbehandlereIEnhet] = useApiCall(saksbehandlereIEnhetApi)
-  const [saksbehandlereForEnhet, setSaksbehandlereForEnhet] = useState<Array<Saksbehandler>>([])
 
   const location = useLocation()
   const navigate = useNavigate()
 
-  useEffect(() => {
-    if (location.pathname.includes('minoppgaveliste')) {
-      if (oppgaveListeValg !== 'MinOppgaveliste') {
-        setOppgaveListeValg('MinOppgaveliste')
-      }
-    }
-  }, [location.pathname])
+  const [oppgavelistaOppgaver, setOppgavelistaOppgaver] = useState<Array<OppgaveDTO>>([])
+  const [minOppgavelisteOppgaver, setMinOppgavelisteOppgaver] = useState<Array<OppgaveDTO>>([])
+
+  const [oppgavelistaFilter, setOppgavelistaFilter] = useState<Filter>(hentFilterFraLocalStorage())
+  const [minOppgavelisteFilter, setMinOppgavelisteFilter] = useState<Filter>(minOppgavelisteFiltre())
+
+  const [oppgavelistaOppgaverResult, hentOppgavelistaOppgaverFetch] = useApiCall(hentOppgaverMedStatus)
+  const [minOppgavelisteOppgaverResult, hentMinOppgavelisteOppgaverFetch] = useApiCall(hentOppgaverMedStatus)
+  const [gosysOppgaverResult, hentGosysOppgaverFetch] = useApiCall(hentGosysOppgaver)
+
+  const [, hentSaksbehandlereIEnheterFetch] = useApiCall(saksbehandlereIEnhetApi)
+  const [saksbehandlereIEnheter, setSaksbehandlereIEnheter] = useState<Array<Saksbehandler>>([])
 
   const oppdaterOppgavelisteValg = (oppgaveListeValg: OppgavelisteToggle) => {
     setOppgaveListeValg(oppgaveListeValg)
@@ -51,83 +58,32 @@ export const ToggleMinOppgaveliste = () => {
     }
   }
 
-  const [minsideFilter, setMinsideFilter] = useState<Filter>(minOppgavelisteFiltre())
-  const [hovedsideFilter, setHovedsideFilter] = useState<Filter>(hentFilterFraLocalStorage())
-
-  useEffect(() => {
-    leggFilterILocalStorage(hovedsideFilter)
-  }, [hovedsideFilter])
-
-  const [minsideOppgaverResult, hentOppgaverMinside] = useApiCall(hentOppgaverMedStatus)
-  const [hovedsideOppgaverResult, hentAlleOppgaverStatusFetch] = useApiCall(hentOppgaverMedStatus)
-  const [gosysOppgaverResult, hentGosysOppgaverFunc] = useApiCall(hentGosysOppgaver)
-
-  const hentMinsideOppgaver = (oppgavestatusFilter: Array<string> | undefined) =>
-    hentOppgaverMinside({
-      oppgavestatusFilter: oppgavestatusFilter ? oppgavestatusFilter : minsideFilter.oppgavestatusFilter,
+  const hentMinOppgavelisteOppgaver = (oppgavestatusFilter?: Array<string>) =>
+    hentMinOppgavelisteOppgaverFetch({
+      oppgavestatusFilter: oppgavestatusFilter ? oppgavestatusFilter : minOppgavelisteFilter.oppgavestatusFilter,
       minOppgavelisteIdent: true,
     })
 
-  const hentHovedsideOppgaver = (oppgavestatusFilter: Array<string> | undefined) =>
-    hentAlleOppgaverStatusFetch({
-      oppgavestatusFilter: oppgavestatusFilter ? oppgavestatusFilter : hovedsideFilter.oppgavestatusFilter,
+  const hentOppgavelistaOppgaver = (oppgavestatusFilter?: Array<string>) =>
+    hentOppgavelistaOppgaverFetch({
+      oppgavestatusFilter: oppgavestatusFilter ? oppgavestatusFilter : oppgavelistaFilter.oppgavestatusFilter,
       minOppgavelisteIdent: false,
     })
 
-  const hentHovedsideOppgaverAlle = () => {
-    hentMinsideOppgaver(undefined)
-    hentGosysOppgaverFunc({})
+  const hentAlleMinOppgavelisteOppgaver = () => {
+    hentMinOppgavelisteOppgaver()
+    hentGosysOppgaverFetch({})
   }
 
   const hentAlleOppgaver = () => {
-    hentMinsideOppgaver(undefined)
-    hentHovedsideOppgaver(undefined)
-    hentGosysOppgaverFunc({})
+    hentMinOppgavelisteOppgaver()
+    hentOppgavelistaOppgaver()
+    hentGosysOppgaverFetch({})
   }
-
-  useEffect(() => {
-    hentAlleOppgaver()
-    if (!!innloggetSaksbehandler.enheter.length) {
-      hentSaksbehandlereIEnhet({ enheter: innloggetSaksbehandler.enheter }, (saksbehandlere) => {
-        setSaksbehandlereForEnhet(saksbehandlere)
-      })
-    }
-  }, [])
 
   const filtrerKunInnloggetBrukerOppgaver = (oppgaver: Array<OppgaveDTO>) => {
     return oppgaver.filter((o) => o.saksbehandlerIdent === innloggetSaksbehandler.ident)
   }
-
-  const [hovedsideOppgaver, setHovedsideOppgaver] = useState<Array<OppgaveDTO>>([])
-  const [minsideOppgaver, setMinsideOppgaver] = useState<Array<OppgaveDTO>>([])
-
-  useEffect(() => {
-    if (isSuccess(hovedsideOppgaverResult) && isSuccess(gosysOppgaverResult)) {
-      const alleOppgaverMerget = sorterOppgaverEtterOpprettet([
-        ...hovedsideOppgaverResult.data,
-        ...gosysOppgaverResult.data,
-      ])
-      setHovedsideOppgaver(alleOppgaverMerget)
-    } else if (isSuccess(hovedsideOppgaverResult) && !isSuccess(gosysOppgaverResult)) {
-      setHovedsideOppgaver(sorterOppgaverEtterOpprettet(hovedsideOppgaverResult.data))
-    } else if (!isSuccess(hovedsideOppgaverResult) && isSuccess(gosysOppgaverResult)) {
-      setHovedsideOppgaver(sorterOppgaverEtterOpprettet(gosysOppgaverResult.data))
-    }
-  }, [hovedsideOppgaverResult, gosysOppgaverResult])
-
-  useEffect(() => {
-    if (isSuccess(minsideOppgaverResult) && isSuccess(gosysOppgaverResult)) {
-      const alleOppgaverMerget = sorterOppgaverEtterOpprettet([
-        ...minsideOppgaverResult.data,
-        ...filtrerKunInnloggetBrukerOppgaver(gosysOppgaverResult.data),
-      ])
-      setMinsideOppgaver(alleOppgaverMerget)
-    } else if (isSuccess(minsideOppgaverResult) && !isSuccess(gosysOppgaverResult)) {
-      setMinsideOppgaver(sorterOppgaverEtterOpprettet(minsideOppgaverResult.data))
-    } else if (!isSuccess(minsideOppgaverResult) && isSuccess(gosysOppgaverResult)) {
-      setMinsideOppgaver(sorterOppgaverEtterOpprettet(filtrerKunInnloggetBrukerOppgaver(gosysOppgaverResult.data)))
-    }
-  }, [gosysOppgaverResult, minsideOppgaverResult])
 
   const oppdaterSaksbehandlerTildeling = (
     oppgave: OppgaveDTO,
@@ -135,18 +91,72 @@ export const ToggleMinOppgaveliste = () => {
     versjon: number | null
   ) => {
     setTimeout(() => {
-      setHovedsideOppgaver(finnOgOppdaterSaksbehandlerTildeling(hovedsideOppgaver, oppgave.id, saksbehandler, versjon))
+      setOppgavelistaOppgaver(
+        finnOgOppdaterSaksbehandlerTildeling(oppgavelistaOppgaver, oppgave.id, saksbehandler, versjon)
+      )
       if (innloggetSaksbehandler.ident === saksbehandler) {
-        setMinsideOppgaver(leggTilOppgavenIMinliste(minsideOppgaver, oppgave, saksbehandler, versjon))
+        setMinOppgavelisteOppgaver(leggTilOppgavenIMinliste(minOppgavelisteOppgaver, oppgave, saksbehandler, versjon))
       } else {
-        setMinsideOppgaver(
+        setMinOppgavelisteOppgaver(
           filtrerKunInnloggetBrukerOppgaver(
-            finnOgOppdaterSaksbehandlerTildeling(minsideOppgaver, oppgave.id, saksbehandler, versjon)
+            finnOgOppdaterSaksbehandlerTildeling(minOppgavelisteOppgaver, oppgave.id, saksbehandler, versjon)
           )
         )
       }
     }, 2000)
   }
+
+  useEffect(() => {
+    if (location.pathname.includes('minoppgaveliste')) {
+      if (oppgaveListeValg !== 'MinOppgaveliste') {
+        setOppgaveListeValg('MinOppgaveliste')
+      }
+    }
+  }, [location.pathname])
+
+  useEffect(() => {
+    leggFilterILocalStorage(oppgavelistaFilter)
+  }, [oppgavelistaFilter])
+
+  useEffect(() => {
+    hentAlleOppgaver()
+    if (!!innloggetSaksbehandler.enheter.length) {
+      hentSaksbehandlereIEnheterFetch({ enheter: innloggetSaksbehandler.enheter }, (saksbehandlere) => {
+        setSaksbehandlereIEnheter(saksbehandlere)
+      })
+    }
+  }, [])
+
+  // Denne spaghettien er pga at vi må ha Gosys oppgaver inn i våres oppgave behandling
+  useEffect(() => {
+    if (isSuccess(oppgavelistaOppgaverResult) && isSuccess(gosysOppgaverResult)) {
+      const alleOppgaverMerget = sorterOppgaverEtterOpprettet([
+        ...oppgavelistaOppgaverResult.data,
+        ...gosysOppgaverResult.data,
+      ])
+      setOppgavelistaOppgaver(alleOppgaverMerget)
+    } else if (isSuccess(oppgavelistaOppgaverResult) && !isSuccess(gosysOppgaverResult)) {
+      setOppgavelistaOppgaver(sorterOppgaverEtterOpprettet(oppgavelistaOppgaverResult.data))
+    } else if (!isSuccess(oppgavelistaOppgaverResult) && isSuccess(gosysOppgaverResult)) {
+      setOppgavelistaOppgaver(sorterOppgaverEtterOpprettet(gosysOppgaverResult.data))
+    }
+  }, [oppgavelistaOppgaverResult, gosysOppgaverResult])
+
+  useEffect(() => {
+    if (isSuccess(minOppgavelisteOppgaverResult) && isSuccess(gosysOppgaverResult)) {
+      const alleOppgaverMerget = sorterOppgaverEtterOpprettet([
+        ...minOppgavelisteOppgaverResult.data,
+        ...filtrerKunInnloggetBrukerOppgaver(gosysOppgaverResult.data),
+      ])
+      setMinOppgavelisteOppgaver(alleOppgaverMerget)
+    } else if (isSuccess(minOppgavelisteOppgaverResult) && !isSuccess(gosysOppgaverResult)) {
+      setMinOppgavelisteOppgaver(sorterOppgaverEtterOpprettet(minOppgavelisteOppgaverResult.data))
+    } else if (!isSuccess(minOppgavelisteOppgaverResult) && isSuccess(gosysOppgaverResult)) {
+      setMinOppgavelisteOppgaver(
+        sorterOppgaverEtterOpprettet(filtrerKunInnloggetBrukerOppgaver(gosysOppgaverResult.data))
+      )
+    }
+  }, [gosysOppgaverResult, minOppgavelisteOppgaverResult])
 
   return (
     <Container>
@@ -157,38 +167,42 @@ export const ToggleMinOppgaveliste = () => {
         }}
       >
         <Tabs.List>
-          <Tabs.Tab value="Oppgavelista" label={`Oppgavelisten (${hovedsideOppgaver.length})`} icon={<InboxIcon />} />
+          <Tabs.Tab
+            value="Oppgavelista"
+            label={`Oppgavelisten (${oppgavelistaOppgaver.length})`}
+            icon={<InboxIcon />}
+          />
           <Tabs.Tab
             value="MinOppgaveliste"
-            label={`Min oppgaveliste (${minsideOppgaver.length})`}
+            label={`Min oppgaveliste (${minOppgavelisteOppgaver.length})`}
             icon={<PersonIcon aria-hidden />}
           />
         </Tabs.List>
       </TabsWidth>
       {oppgaveListeValg === 'MinOppgaveliste' ? (
         <MinOppgaveliste
-          minsideOppgaver={minsideOppgaver}
-          minsideOppgaverResult={minsideOppgaverResult}
+          minsideOppgaver={minOppgavelisteOppgaver}
+          minsideOppgaverResult={minOppgavelisteOppgaverResult}
           gosysOppgaverResult={gosysOppgaverResult}
-          minsideFilter={minsideFilter}
+          minsideFilter={minOppgavelisteFilter}
           setMinsideFilter={(filter: Filter) => {
-            hentMinsideOppgaver(filter.oppgavestatusFilter)
-            setMinsideFilter(filter)
+            hentMinOppgavelisteOppgaver(filter.oppgavestatusFilter)
+            setMinOppgavelisteFilter(filter)
           }}
-          setMinsideOppgaver={setMinsideOppgaver}
-          saksbehandlereIEnhet={saksbehandlereForEnhet}
+          setMinsideOppgaver={setMinOppgavelisteOppgaver}
+          saksbehandlereIEnhet={saksbehandlereIEnheter}
           oppdaterSaksbehandlerTildeling={oppdaterSaksbehandlerTildeling}
         />
       ) : (
-        <OppgavelistaWrapper
-          saksbehandlereIEnhet={saksbehandlereForEnhet}
-          hovedsideOppgaver={hovedsideOppgaver}
-          hentHovedsideOppgaverAlle={hentHovedsideOppgaverAlle}
-          hovedsideOppgaverResult={hovedsideOppgaverResult}
+        <Oppgavelista
+          saksbehandlereIEnhet={saksbehandlereIEnheter}
+          hovedsideOppgaver={oppgavelistaOppgaver}
+          hentHovedsideOppgaverAlle={hentAlleMinOppgavelisteOppgaver}
+          hovedsideOppgaverResult={oppgavelistaOppgaverResult}
           gosysOppgaverResult={gosysOppgaverResult}
-          hentHovedsideOppgaver={hentHovedsideOppgaver}
-          hovedsideFilter={hovedsideFilter}
-          setHovedsideFilter={setHovedsideFilter}
+          hentHovedsideOppgaver={hentOppgavelistaOppgaver}
+          hovedsideFilter={oppgavelistaFilter}
+          setHovedsideFilter={setOppgavelistaFilter}
           oppdaterSaksbehandlerTildeling={oppdaterSaksbehandlerTildeling}
         />
       )}
