+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/oppgaveFiltrering/VelgOppgavestatuser.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { UNSAFE_Combobox } from '@navikt/ds-react'
-import { OPPGAVESTATUSFILTER } from '~components/oppgavebenk/filter/oppgavelistafiltre'
+import { OPPGAVESTATUSFILTER } from '~components/oppgavebenk/oppgaveFiltrering/oppgavelistafiltre'
 
 interface Props {
   value: Array<string>
