+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/oppgaverTable/OppgaverTableRow.tsx
@@ -2,13 +2,13 @@ import React, { ReactNode } from 'react'
 import { Table } from '@navikt/ds-react'
 import { erOppgaveRedigerbar, OppgaveDTO } from '~shared/api/oppgaver'
 import { formaterStringDato } from '~utils/formattering'
-import { FristWrapper } from '~components/oppgavebenk/FristWrapper'
-import SaksoversiktLenke from '~components/oppgavebenk/SaksoversiktLenke'
-import { OppgavetypeTag, SaktypeTag } from '~components/oppgavebenk/Tags'
-import { HandlingerForOppgave } from '~components/oppgavebenk/HandlingerForOppgave'
-import { FristHandlinger } from '~components/oppgavebenk/FristHandlinger'
+import { FristWrapper } from '~components/oppgavebenk/frist/FristWrapper'
+import SaksoversiktLenke from '~components/oppgavebenk/components/SaksoversiktLenke'
+import { OppgavetypeTag, SaktypeTag } from '~components/oppgavebenk/components/Tags'
+import { HandlingerForOppgave } from '~components/oppgavebenk/components/HandlingerForOppgave'
+import { FristHandlinger } from '~components/oppgavebenk/frist/FristHandlinger'
 import { VelgSaksbehandler } from '~components/oppgavebenk/tildeling/VelgSaksbehandler'
-import { OPPGAVESTATUSFILTER } from '~components/oppgavebenk/filter/oppgavelistafiltre'
+import { OPPGAVESTATUSFILTER } from '~components/oppgavebenk/oppgaveFiltrering/oppgavelistafiltre'
 import { Saksbehandler } from '~shared/types/saksbehandler'
 
 interface Props {
