+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/utils/oppgaveutils.ts
