+++ apps/etterlatte-saksbehandling-ui/client/src/components/person/journalfoeringsoppgave/handling/StartOppgavebehandling.tsx
@@ -8,8 +8,7 @@ import { formaterOppgaveStatus, formaterSakstype, formaterStringDato } from '~ut
 import { InfoWrapper } from '~components/behandling/soeknadsoversikt/styled'
 import { Info } from '~components/behandling/soeknadsoversikt/Info'
 import { FlexRow } from '~shared/styled'
-
-import { FristWrapper } from '~components/oppgavebenk/FristWrapper'
+import { FristWrapper } from '~components/oppgavebenk/frist/FristWrapper'
 import { OppgaveHandling, settOppgaveHandling } from '~store/reducers/JournalfoeringOppgaveReducer'
 import { FormWrapper } from '../BehandleJournalfoeringOppgave'
 import { useFeatureEnabledMedDefault } from '~shared/hooks/useFeatureToggle'
